+++ CHANGELOG.md
@@ -20,11 +20,13 @@ The API consists of all public Java types from `com.atlassian.performance.tools.
 
 ### Fixed
 - Download Jira installer from the official downloads site. Resolve [JPERF-277].
+- Use StackVirtualUsersFormula instead of Ec2VirtualUsersFormula. Work around [JPERF-406].
 
 ### Added
 - Add support for `aws-infrastructure:2.6.0`.
 
 [JPERF-277]: https://ecosystem.atlassian.net/browse/JPERF-277
+[JPERF-406]: https://ecosystem.atlassian.net/browse/JPERF-406
 
 ## [3.1.0] - 2019-02-19
 [3.1.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-3.1.0%0Drelease-3.0.0